Update patches so every modules builds against the same Windows SDK
A lot of the older modules still hardcode targeting older Windows SDK versions. If you're setting up a new dev machine and follow our current guidance (aka using the RNW dependency script) you won't have the older SDKs needed to build Gallery.

We've partially addressed this with some minor patch files in this repo, (which set the modules to use newer SDK versions), but this PR updates the patches in a more future-proof way so the libraries are all building on the same Windows SDK that RNW and the app are using.